--- src/sidebar/FeedList/index.ts
@@ -1,3 +1,0 @@
-import FeedList from './FeedList';
-
-export default FeedList;
