+++ src/sidebar/MainView/folder-tree/Folder/Folder.tsx
@@ -2,12 +2,12 @@ import { FolderSimple, CaretDown, CaretRight } from 'phosphor-react';
 
 import React, { Fragment, useMemo, useRef, useState } from 'react';
 
-import { TreeNode } from '../../../model/feeds';
-import { useAppDispatch, useAppSelector } from '../../../store/hooks';
-import feedsSlice, { selectHasVisibleChildren } from '../../../store/slices/feeds';
-import { RelativeDragDropPosition } from '../../../utils/dragdrop';
-import { MouseEventButton } from '../../../utils/types/web-api';
-import { useContextMenu } from '../../Menu/ContextMenu/useContextMenu';
+import { TreeNode } from '../../../../model/feeds';
+import { useAppDispatch, useAppSelector } from '../../../../store/hooks';
+import feedsSlice, { selectHasVisibleChildren } from '../../../../store/slices/feeds';
+import { RelativeDragDropPosition } from '../../../../utils/dragdrop';
+import { MouseEventButton } from '../../../../utils/types/web-api';
+import { useContextMenu } from '../../../Menu/ContextMenu/useContextMenu';
 import useDragDropNode from '../dragdrop/useDragDropNode';
 import { clsx } from 'clsx';
