+++ src/sidebar/MainView/folder-tree/FolderTreeNode.tsx
@@ -1,7 +1,7 @@
 import React, { Fragment, memo, useMemo } from 'react';
 
-import { useAppSelector } from '../../store/hooks';
-import { makeSelectTreeNode } from '../../store/slices/feeds';
+import { useAppSelector } from '../../../store/hooks';
+import { makeSelectTreeNode } from '../../../store/slices/feeds';
 import Folder from './Folder/Folder';
 import FolderSubTreeNode from './FolderSubTreeNode';
