+++ src/sidebar/MainView/folder-tree/folder-tree.css
@@ -1,0 +1,2 @@
+@import './Folder/folder.css';
+@import './FolderEdit/folder-edit.css';
