+++ src/sidebar/MainView/main-view.css
@@ -1,0 +1,2 @@
+@import './folder-tree/folder-tree.css';
+@import './FeedList/feed-list.css';
